+++ ios/Classes/CleverTapPlugin.h
@@ -1,12 +1,14 @@
 #import <Flutter/Flutter.h>
 
-static NSString *const kCleverTapProfileDidInitialize       = @"profileDidInitialize";
-static NSString *const kCleverTapProfileSync                = @"profileDataUpdated";
-static NSString *const kCleverTapInAppNotificationDismissed = @"inAppNotificationDismissed";
-static NSString *const kCleverTapInboxDidInitialize         = @"inboxDidInitialize";
-static NSString *const kCleverTapInboxMessagesDidUpdate     = @"inboxMessagesDidUpdate";
-static NSString *const kCleverTapExperimentsDidUpdate       = @"CTExperimentsUpdated";
-static NSString *const kCleverTapDisplayUnitsLoaded         = @"onDisplayUnitsLoaded";
+static NSString *const kCleverTapProfileDidInitialize             = @"profileDidInitialize";
+static NSString *const kCleverTapProfileSync                      = @"profileDataUpdated";
+static NSString *const kCleverTapInAppNotificationDismissed       = @"inAppNotificationDismissed";
+static NSString *const kCleverTapInboxDidInitialize               = @"inboxDidInitialize";
+static NSString *const kCleverTapInboxMessagesDidUpdate           = @"inboxMessagesDidUpdate";
+static NSString *const kCleverTapExperimentsDidUpdate             = @"CTExperimentsUpdated";
+static NSString *const kCleverTapDisplayUnitsLoaded               = @"onDisplayUnitsLoaded";
+static NSString *const kCleverTapInAppNotificationButtonTapped    = @"onInAppButtonClick";
+static NSString *const kCleverTapInboxMessageButtonTapped         = @"onInboxButtonClick";
 
 @interface CleverTapPlugin : NSObject<FlutterPlugin>
 + (instancetype)sharedInstance;