+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## CHANGE LOG
 
+### Version 1.1.1 (March 30, 2020)
+* Adds support for Custom App Inbox
+* Adds support for InApp/Inbox button click listeners
+* Adds support for Notification Clicked/Viewed for App Inbox
+* Adds support for passing Xiaomi/Baidu tokens.
+* Supports CleverTap Android SDK v3.7.2
+* Supports CleverTap iOS SDK v3.7.3
+
 ### Version 1.1.0 (February 27, 2020)
 * Adds support for Dynamic Variables & Native Display
 * Adds support for Google Play Install Referrer Library v1.0